--- packages/broker/src/errors/FailedToPublishError.ts
@@ -1,5 +1,0 @@
-export class FailedToPublishError extends Error {
-    constructor(streamId: string, reason: string) {
-        super(`Failed publish to stream ${streamId}, reason: ${reason}`)
-    }
-}
