--- packages/broker/src/errors/GenericError.ts
@@ -1,9 +1,0 @@
-export class GenericError extends Error {
-
-    code: string
-
-    constructor(message: string, code: string) {
-        super(message)
-        this.code = code
-    }
-}
