+++ packages/broker/src/plugins/websocket/WebsocketPlugin.ts
@@ -1,9 +1,13 @@
 import { Plugin, PluginOptions } from '../../Plugin'
-import { SslCertificateConfig } from '../../types'
 import { getPayloadFormat } from '../../helpers/PayloadFormat'
 import { WebsocketServer } from './WebsocketServer'
 import PLUGIN_CONFIG_SCHEMA from './config.schema.json'
 import { Schema } from 'ajv'
+
+export interface SslCertificateConfig {
+    privateKeyFileName: string
+    certFileName: string
+}
 
 export interface WebsocketPluginConfig {
     port: number
