+++ packages/broker/src/plugins/websocket/WebsocketServer.ts
@@ -10,7 +10,7 @@ import StreamrClient from 'streamr-client'
 import { Logger } from 'streamr-network'
 import { Connection } from './Connection'
 import { ApiAuthenticator } from '../../apiAuthenticator'
-import { SslCertificateConfig } from '../../types'
+import { SslCertificateConfig } from './WebsocketPlugin'
 import { PublishConnection } from './PublishConnection'
 import { SubscribeConnection } from './SubscribeConnection'
 import { PayloadFormat } from '../../helpers/PayloadFormat'
