--- packages/broker/src/types.ts
@@ -1,4 +1,0 @@
-export interface SslCertificateConfig {
-    privateKeyFileName: string
-    certFileName: string
-}
